requirements.txt
argcomplete==1.9.2
atomicwrites==1.1.5
attrs==18.1.0
base58==0.2.4
boto3==1.4.7
botocore==1.7.14
certifi==2018.4.16
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
colorama==0.3.9
diff-match-patch==20121119
Django==2.0.5
django-import-export==1.0.1
djangorestframework==3.8.2
docutils==0.14
durationpy==0.5
et-xmlfile==1.0.1
future==0.16.0
futures==3.1.1
hjson==3.0.0
idna==2.6
jdcal==1.4
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.16.1
more-itertools==4.2.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.3
placebo==0.8.1
pluggy==0.6.0
py==1.5.3
pytest==3.6.0
pytest-django==3.2.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.4
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
tablib==0.12.1
toml==0.9.2
tqdm==4.15.0
troposphere==2.3.0
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.12
wsgi-request-logger==0.4.6
xlrd==1.1.0
xlwt==1.3.0
zappa==0.44.3
